abstracted functions
