+++ discordStuff.js
@@ -5,10 +5,8 @@ const {
   MessageActionRow,
   MessageButton,
 } = require("discord.js");
-let unisJson = require("./unis.json");
-const uniData = require("./mongoDB/model");
-const editJsonFile = require("edit-json-file");
-const { channel } = require("diagnostics_channel");
+
+const { getUnisList, changeUniStatus, getUnisArray } = require('./helper/discordFunctions');
 require("dotenv").config({ path: "./.env" });
 
 global.arr = [];
@@ -31,11 +29,10 @@ client.on("interactionCreate", async (interaction) => {
   if (!interaction.isButton()) return;
 
   if (interaction.customId == "accept") {
-    console.log(interaction);
     global.name = interaction.user.username;
 
     let ans = await getUnisList(global.name);
-    global.arr = await getUnisArr(global.name);
+    global.arr = await getUnisArray(global.name);
 
     interaction.update({
       content: `Which of these programs did you get accepted to?\n${ans}`,
@@ -44,7 +41,6 @@ client.on("interactionCreate", async (interaction) => {
   }
 
   if (interaction.customId == "chance") {
-
     interaction.update({
       content: "You have a 100% chance of getting into your top choice!",
       components: [],
@@ -54,10 +50,8 @@ client.on("interactionCreate", async (interaction) => {
 
 client.on("messageCreate", async (msg) => {
   if (msg.content == "!show") {
-    if (
-      msg.channelId == "933150268852408330" ||
-      msg.channelId == "933473094968946768"
-    ) {
+    if (msg.channelId == process.env.UNIBOT_THREAD || msg.channelId == process.env.BOT_TESTING) {
+      
       let name = msg.author.username;
       let uniList = await getUnisList(name);
 
@@ -108,70 +102,3 @@ client.on("messageCreate", async (msg) => {
 });
 
 client.login(process.env.DISCORD_TOKEN);
-
-// function addPerson(){
-//     let file = editJsonFile(`${__dirname}/unis.json`);
-//     file.set(`${name}.${uni}`, "Accepted");
-//     file.save();
-
-//     delete require.cache[require.resolve("./unis.json")];
-//     unisJson = require("./unis.json");
-// }
-
-function changeUniStatus(nameRec, uni) {
-  uniData.findOne({ name: nameRec }, function (err, res) {
-    let personUniList = res.unis;
-    personUniList[uni] = "Accepted";
-    console.log(personUniList);
-    uniData.findOneAndUpdate(
-      { name: nameRec },
-      { unis: personUniList },
-      { returnOriginal: false },
-      function (err, doc) {
-        if (err) {
-          console.log("Something went wrong with the upload");
-        }
-
-        // console.log(doc);
-      }
-    );
-  });
-}
-
-async function getUnisList(nameRec) {
-  try {
-    let list = await uniData
-      .findOne({ name: nameRec }, function (err, res) {})
-      .clone();
-
-    let ans = "**";
-    list = list.unis;
-
-    for (var i in list) {
-      if (list[i] == "Accepted") {
-        ans += i + "✅";
-        ans += "\n";
-      } else {
-        ans += i;
-        ans += "\n";
-      }
-    }
-    ans += "**";
-    return ans;
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-async function getUnisArr(name) {
-  let ans = [];
-
-  let list = await uniData.findOne({ name: name }).clone();
-  list = list.unis;
-
-  for (var i in list) {
-    ans.push(i);
-  }
-
-  return ans;
-}
